feat: 안전하게 사진 삭제하도록 수정

+++ sum3years-app/src/main/java/com/sum3years/ui/main/DownloadedScreen.kt
@@ -79,8 +79,13 @@ fun DownloadedScreen() {
             },
             confirmButton = {
                 TextButton(onClick = {
-                    file.delete()
-                    deleteTarget = null
+                    if (file.exists()) {
+                        try {
+                            file.delete()
+                        } finally {
+                            deleteTarget = null
+                        }
+                    }
                 }) {
                     Text("삭제")
                 }
